+++ Source/ImGui/Private/ImGuiImplementation.cpp
@@ -32,7 +32,6 @@ ImGuiContext** GImGuiContextPtrHandle = &GImGuiContextPtr;
 #include "imgui_demo.cpp"
 #include "imgui_draw.cpp"
 #include "imgui_widgets.cpp"
-#include "misc/stl/imgui_stl.cpp"
 
 #if PLATFORM_WINDOWS
 #include <Windows/HideWindowsPlatformTypes.h>