+++ Source/ImGui/Private/ImGuiInteroperability.cpp
@@ -7,8 +7,44 @@
 #include "Utilities/Arrays.h"
 
 
+// If TCHAR is wider than ImWchar, enable or disable validation of input character before conversions.
+#define VALIDATE_INPUT_CHARACTERS 1
+
+#if VALIDATE_INPUT_CHARACTERS
+DEFINE_LOG_CATEGORY_STATIC(LogImGuiInput, Warning, All);
+#endif
+
 namespace
 {
+	//====================================================================================================
+	// Character conversion
+	//====================================================================================================
+
+	template<typename T, std::enable_if_t<(sizeof(T) <= sizeof(ImWchar)), T>* = nullptr>
+	ImWchar CastInputChar(T Char)
+	{
+		return static_cast<ImWchar>(Char);
+	}
+
+	template<typename T, std::enable_if_t<!(sizeof(T) <= sizeof(ImWchar)), T>* = nullptr>
+	ImWchar CastInputChar(T Char)
+	{
+#if VALIDATE_INPUT_CHARACTERS
+		// We only need a runtime validation if TCHAR is wider than ImWchar.
+		// Signed and unsigned integral types with the same size as ImWchar should be safely converted. As long as the
+		// char value is in that range we can safely use it, otherwise we should log an error to notify about possible
+		// truncations.
+		static constexpr auto MinLimit = (std::numeric_limits<std::make_signed_t<ImWchar>>::min)();
+		static constexpr auto MaxLimit = (std::numeric_limits<std::make_unsigned_t<ImWchar>>::max)();
+		UE_CLOG(!(Char >= MinLimit && Char <= MaxLimit), LogImGuiInput, Error,
+			TEXT("TCHAR value '%c' (%#x) is out of range %d (%#x) to %u (%#x) that can be safely converted to ImWchar. ")
+			TEXT("If you wish to disable this validation, please set VALIDATE_INPUT_CHARACTERS in ImGuiInputState.cpp to 0."),
+			Char, Char, MinLimit, MinLimit, MaxLimit, MaxLimit);
+#endif
+
+		return static_cast<ImWchar>(Char);
+	}
+
 	//====================================================================================================
 	// Copying Utilities
 	//====================================================================================================
@@ -247,16 +283,6 @@ namespace ImGuiInterops
 		static const uint32 LeftAlt = GetKeyIndex(EKeys::LeftAlt);
 		static const uint32 RightAlt = GetKeyIndex(EKeys::RightAlt);
 
-		// Check whether we need to draw cursor.
-		IO.MouseDrawCursor = InputState.HasMousePointer();
-
-		// Copy mouse position.
-		IO.MousePos.x = InputState.GetMousePosition().X;
-		IO.MousePos.y = InputState.GetMousePosition().Y;
-
-		// Copy mouse wheel delta.
-		IO.MouseWheel += InputState.GetMouseWheelDelta();
-
 		// Copy key modifiers.
 		IO.KeyCtrl = InputState.IsControlDown();
 		IO.KeyShift = InputState.IsShiftDown();
@@ -274,9 +300,9 @@ namespace ImGuiInterops
 			Copy(InputState.GetMouseButtons(), IO.MouseDown, InputState.GetMouseButtonsUpdateRange());
 		}
 
-		if (InputState.GetCharactersNum() > 0)
+		for (const TCHAR Char : InputState.GetCharacters())
 		{
-			Copy(InputState.GetCharacters(), IO.InputCharacters);
+			IO.AddInputCharacter(CastInputChar(Char));
 		}
 
 		if (InputState.IsGamepadNavigationEnabled() && InputState.HasGamepad())
@@ -287,5 +313,28 @@ namespace ImGuiInterops
 		SetFlag(IO.ConfigFlags, ImGuiConfigFlags_NavEnableKeyboard, InputState.IsKeyboardNavigationEnabled());
 		SetFlag(IO.ConfigFlags, ImGuiConfigFlags_NavEnableGamepad, InputState.IsGamepadNavigationEnabled());
 		SetFlag(IO.BackendFlags, ImGuiBackendFlags_HasGamepad, InputState.HasGamepad());
+
+		// Check whether we need to draw cursor.
+		IO.MouseDrawCursor = InputState.HasMousePointer();
+
+		// If touch is enabled and active, give it a precedence.
+		if (InputState.IsTouchActive())
+		{
+			// Copy the touch position to mouse position.
+			IO.MousePos.x = InputState.GetTouchPosition().X;
+			IO.MousePos.y = InputState.GetTouchPosition().Y;
+
+			// With touch active one frame longer than it is down, we have one frame to processed touch up.
+			IO.MouseDown[0] = InputState.IsTouchDown();
+		}
+		else
+		{
+			// Copy the mouse position.
+			IO.MousePos.x = InputState.GetMousePosition().X;
+			IO.MousePos.y = InputState.GetMousePosition().Y;
+
+			// Copy mouse wheel delta.
+			IO.MouseWheel += InputState.GetMouseWheelDelta();
+		}
 	}
 }